refactor: upgrade reference id from uint to ulong
Widen the reference ID type from uint (32-bit, ~4.3B) to ulong (64-bit)
across the engine serialization layer. This eliminates the theoretical
wraparound exhaustion risk for long-running sandbox saves where entity
churn (map refresh, automation, combat) permanently leaks IDs.

Engine layer changes:
- ReferenceContext: _nextId, dictionaries, GetId/SetReference/TryGetReference
- ReferenceContextExtension: TryReadReferenceId/TryWriteReferenceId
- BinarySerializeWriteNode: BindReference implementation
- BinaryPostLoadSerializeNode: TryGetId/BindReference

Breaking: existing save files are incompatible ($id wire size 4 -> 8 bytes).

+++ Src/Alco/Serialization/Binary/BinaryPostLoadSerializeNode.cs
@@ -310,7 +310,7 @@ public class BinaryPostLoadSerializeNode : SerializeNode
         if (_clearMissingReferences)
             referenceable = default;
 
-        if (!TryGetId(key, out uint id) || id == 0)
+        if (!TryGetId(key, out ulong id) || id == 0)
             return;
 
         if (_referenceContext.TryGetReference(id, out object? obj) && obj is T reference)
@@ -322,9 +322,9 @@ public class BinaryPostLoadSerializeNode : SerializeNode
         AddError($"Failed to resolve reference '{key}': {id}\n{Environment.StackTrace}");
     }
 
-    private bool TryGetId(string key, out uint id)
+    private bool TryGetId(string key, out ulong id)
     {
-        if (_content.TryGetValue(key, out uint v))
+        if (_content.TryGetValue(key, out ulong v))
         {
             id = v;
             return true;
